pyproject.toml
[tool.poetry]
name = "dawa-pahchan"
version = "2.0.0"
description = "دوا پہچان - Medicine Recognition & Safety PWA for Pakistan"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.13"
fastapi = ">=0.104.0"
uvicorn = ">=0.24.0"
python-multipart = ">=0.0.6"
google-generativeai = ">=0.3.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
